Fixed a timer bug where you were capped to 24 days
Because timer was using an integer of milliseconds, the number of days you could set a timer for was limited to 24. I converted this to a long and seemingly fixed the issue.

+++ OvergrowthApp/app/src/main/java/com/example/overgrowthapp/ui/PlantDatabaseHelper.java
@@ -141,7 +141,7 @@ public class PlantDatabaseHelper extends SQLiteOpenHelper {
                 plantPersonal plant = new plantPersonal();
                 plant.setId(cursor.getInt(cursor.getColumnIndex("id")));
                 plant.setTimerEnd(cursor.getLong(cursor.getColumnIndex("timer_end")));
-                plant.setTimerLength(cursor.getInt(cursor.getColumnIndex("timer_length")));
+                plant.setTimerLength(cursor.getLong(cursor.getColumnIndex("timer_length")));
                 plant.setCommonID(cursor.getString(cursor.getColumnIndex("common_id")));
                 plant.setBotanicalID(cursor.getString(cursor.getColumnIndex("botanical_id")));
                 plant.setImgSrc(cursor.getString(cursor.getColumnIndex("img_src")));
@@ -178,7 +178,7 @@ public class PlantDatabaseHelper extends SQLiteOpenHelper {
                 plantPersonal plant = new plantPersonal();
                 plant.setId(cursor.getInt(cursor.getColumnIndex("id")));
                 plant.setTimerEnd(cursor.getLong(cursor.getColumnIndex("timer_end")));
-                plant.setTimerLength(cursor.getInt(cursor.getColumnIndex("timer_length")));
+                plant.setTimerLength(cursor.getLong(cursor.getColumnIndex("timer_length")));
                 plant.setCommonID(cursor.getString(cursor.getColumnIndex("common_id")));
                 plant.setBotanicalID(cursor.getString(cursor.getColumnIndex("botanical_id")));
                 plant.setImgSrc(cursor.getString(cursor.getColumnIndex("img_src")));
